requirements.txt
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
scikit-learn>=1.0.0,<1.4.0
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<0.13.0
pyyaml>=5.4.0,<7.0.0
tabulate>=0.9.0